+++ src/main/java/com/smoothstack/december/administratormicroservice/model/BookLoan.java
@@ -1,4 +1,4 @@
-package com.smoothstack.december.administratormicroservice.models;
+package com.smoothstack.december.administratormicroservice.model;
 
 import java.time.LocalDate;
 import java.util.Objects;;
