+++ src/main/java/com/smoothstack/december/administratormicroservice/model/Borrower.java
@@ -1,10 +1,11 @@
-package com.smoothstack.december.administratormicroservice.models;
+package com.smoothstack.december.administratormicroservice.model;
 
-public class LibraryBranch {
+public class Borrower {
 
     private Long id;
     private String name;
     private String address;
+    private String phoneNumber;
 
     public Long getId() {
         return this.id;
@@ -28,6 +29,14 @@ public class LibraryBranch {
 
     public void setAddress(String address) {
         this.address = address;
+    }
+
+    public String getPhoneNumber() {
+        return this.phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
     }
 
 }
