+++ src/main/java/com/smoothstack/december/administratormicroservice/model/Genre.java
@@ -1,14 +1,12 @@
-package com.smoothstack.december.administratormicroservice.models;
+package com.smoothstack.december.administratormicroservice.model;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public class Publisher {
+public class Genre {
 
     private Long id;
     private String name;
-    private String address;
-    private String phoneNumber;
     private Set<Book> books = new HashSet<>();
 
     public Long getId() {
@@ -25,22 +23,6 @@ public class Publisher {
 
     public void setName(String name) {
         this.name = name;
-    }
-
-    public String getAddress() {
-        return this.address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhoneNumber() {
-        return this.phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
     }
 
     public Set<Book> getBooks() {
