+++ src/main/java/com/smoothstack/december/administratormicroservice/model/Publisher.java
@@ -1,12 +1,14 @@
-package com.smoothstack.december.administratormicroservice.models;
+package com.smoothstack.december.administratormicroservice.model;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public class Genre {
+public class Publisher {
 
     private Long id;
     private String name;
+    private String address;
+    private String phoneNumber;
     private Set<Book> books = new HashSet<>();
 
     public Long getId() {
@@ -23,6 +25,22 @@ public class Genre {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public String getAddress() {
+        return this.address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getPhoneNumber() {
+        return this.phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
     }
 
     public Set<Book> getBooks() {
